pyproject.toml
[tool.poetry]
name = "xsdesignhosting"
version = "0.1.0"
description = ""
authors = ["Pablo Cortés Rodríguez <cortesrodriguezpablo3@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
flask = "^3.0.3"
flask-cors = "^4.0.0"
firebase-admin = "^6.5.0"
pymongo = "^4.7.0"
python-dotenv = "^1.0.1"
pytest = "^8.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
